use influx plugin
