+++ dags/pre-processing.py
@@ -17,6 +17,7 @@ from airflow.contrib.hooks import SSHHook
 from airflow.operators.slack_operator import SlackAPIPostOperator
 from airflow.operators import SlackAPIEnsureChannelOperator, SlackAPIInviteToChannelOperator, SlackAPIUploadFileOperator
 from airflow.operators import FeiEpuOperator
+from airflow.operators import Send2InfluxOperator
 
 from airflow.exceptions import AirflowException, AirflowSkipException
 
@@ -26,14 +27,14 @@ import fileinput
 from pathlib import Path
 import requests
 
-import influxdb
+# import influxdb
 from pprint import pprint, pformat
 
-import json
-from collections import defaultdict, MutableMapping
-import xml.etree.ElementTree as ET
-from ast import literal_eval
-from dateutil import parser
+# import json
+# from collections import defaultdict, MutableMapping
+# import xml.etree.ElementTree as ET
+# from ast import literal_eval
+# from dateutil import parser
 
 import logging
 LOG = logging.getLogger(__name__)
@@ -46,18 +47,18 @@ args = {
 }
 
 
-
-# from https://stackoverflow.com/questions/6027558/flatten-nested-python-dictionaries-compressing-keys
-def flatten(d, parent_key='', sep='.'):
-    """ flatten the dict d so that we can export as key/value pairs """
-    items = []
-    for k, v in d.items():
-        new_key = parent_key + sep + k if parent_key else k
-        if isinstance(v, MutableMapping):
-            items.extend(flatten(v, new_key, sep=sep).items())
-        else:
-            items.append((new_key, v))
-    return dict(items)
+#
+# # from https://stackoverflow.com/questions/6027558/flatten-nested-python-dictionaries-compressing-keys
+# def flatten(d, parent_key='', sep='.'):
+#     """ flatten the dict d so that we can export as key/value pairs """
+#     items = []
+#     for k, v in d.items():
+#         new_key = parent_key + sep + k if parent_key else k
+#         if isinstance(v, MutableMapping):
+#             items.extend(flatten(v, new_key, sep=sep).items())
+#         else:
+#             items.append((new_key, v))
+#     return dict(items)
 
 
 
@@ -67,57 +68,57 @@ def uploadExperimentalParameters2Logbook(ds, **kwargs):
     LOG.warn("data: %s" % (data,))
     raise AirflowSkipException('not yet implemented')
 
-
-class Send2InfluxOperator(PythonOperator):
-    template_fields = ('experiment',)
-    def __init__(self,xcom_task_id,xcom_key,host='localhost',port=8086,user='root',password='root',db='cryoem',measurement='microscope_image',experiment=None,*args,**kwargs):
-        BaseOperator.__init__(self,*args,**kwargs)
-        self.xcom_task_id = xcom_task_id
-        self.xcom_key = xcom_key
-        self.host = host
-        self.port = port
-        self.user = user
-        self.password = password
-        self.db = db
-        self.measurement = measurement
-        self.experiment = experiment
-    def execute(self, context,**kwargs):
-        """Push the parameter key-value pairs to the elogbook"""
-        d = context['ti'].xcom_pull( task_ids=self.xcom_task_id, key=self.xcom_key )['MicroscopeImage']
-        dt = parser.parse( d['microscopeData']['acquisition']['acquisitionDateTime'] )
-        dd = flatten(d, sep='_')
-        context = {}
-        data = {} 
-        for k,v in dd.items():
-            # ignore these entries
-            if k in ( 'microscopeData_acquisition_acquisitionDateTime', 'CustomData_FindFoilHoleCenterResults_@type' ):
-                continue
-            # force context
-            elif k in ( 'microscopeData_instrument_InstrumentID', ):
-                v = '%s' % v
-            # LOG.info("k=%s, v=%s" % (k,v))
-            if isinstance( v, (str,bool) ) or v == None:
-                # LOG.info("  context")
-                vv = "'%s'" % v if isinstance(v,str) and ' ' in v else v
-                context[k] = vv
-            else:
-                # LOG.info("  data %s" % v)
-                data[k] = float(v)
-
-        client = influxdb.InfluxDBClient( self.host, self.port, self.user, self.password, self.db )
-        #LOG.info("DB: %s:%s/%s (%s %s)" % (self.host, self.port, self.db, self.user, self.password) )
-        #LOG.info( '%s @ %s: \n%s\n%s' % (self.measurement,dt.strftime('%s'),pformat(context),pformat(data) ) )
-        client.create_database(self.measurement)
-        if self.experiment:
-           context['experiment'] = self.experiment
-        client.write_points([{
-            "measurement": self.measurement,
-            "tags": context,
-            "fields": data,
-            "time": dt,
-        }])
-        return
-
+#
+# class Send2InfluxOperator(PythonOperator):
+#     template_fields = ('experiment',)
+#     def __init__(self,xcom_task_id,xcom_key,host='localhost',port=8086,user='root',password='root',db='cryoem',measurement='microscope_image',experiment=None,*args,**kwargs):
+#         BaseOperator.__init__(self,*args,**kwargs)
+#         self.xcom_task_id = xcom_task_id
+#         self.xcom_key = xcom_key
+#         self.host = host
+#         self.port = port
+#         self.user = user
+#         self.password = password
+#         self.db = db
+#         self.measurement = measurement
+#         self.experiment = experiment
+#     def execute(self, context,**kwargs):
+#         """Push the parameter key-value pairs to the elogbook"""
+#         d = context['ti'].xcom_pull( task_ids=self.xcom_task_id, key=self.xcom_key )['MicroscopeImage']
+#         dt = parser.parse( d['microscopeData']['acquisition']['acquisitionDateTime'] )
+#         dd = flatten(d, sep='_')
+#         context = {}
+#         data = {}
+#         for k,v in dd.items():
+#             # ignore these entries
+#             if k in ( 'microscopeData_acquisition_acquisitionDateTime', 'CustomData_FindFoilHoleCenterResults_@type' ):
+#                 continue
+#             # force context
+#             elif k in ( 'microscopeData_instrument_InstrumentID', ):
+#                 v = '%s' % v
+#             # LOG.info("k=%s, v=%s" % (k,v))
+#             if isinstance( v, (str,bool) ) or v == None:
+#                 # LOG.info("  context")
+#                 vv = "'%s'" % v if isinstance(v,str) and ' ' in v else v
+#                 context[k] = vv
+#             else:
+#                 # LOG.info("  data %s" % v)
+#                 data[k] = float(v)
+#
+#         client = influxdb.InfluxDBClient( self.host, self.port, self.user, self.password, self.db )
+#         #LOG.info("DB: %s:%s/%s (%s %s)" % (self.host, self.port, self.db, self.user, self.password) )
+#         #LOG.info( '%s @ %s: \n%s\n%s' % (self.measurement,dt.strftime('%s'),pformat(context),pformat(data) ) )
+#         client.create_database(self.measurement)
+#         if self.experiment:
+#            context['experiment'] = self.experiment
+#         client.write_points([{
+#             "measurement": self.measurement,
+#             "tags": context,
+#             "fields": data,
+#             "time": dt,
+#         }])
+#         return
+#
 
 
 
@@ -128,7 +129,7 @@ with DAG( 'cryoem_pre-processing',
         description="Conduct some initial processing to determine efficacy of CryoEM data and upload it to the elogbook",
         schedule_interval=None,
         default_args=args,
-        max_active_runs=3
+        max_active_runs=5
     ) as dag:
 
     # hook to container host for lsf commands
@@ -244,6 +245,9 @@ mrc2tif -j {{ dag_run.conf['directory'] }}/{{ dag_run.conf['base'] }}_ctf.mrc {{
         jobid="{{ ti.xcom_pull( task_ids='ctf_summed' ) }}"
     )
     
+    t_influx_ctf_summed = DummyOperator( task_id='influx_ctf_summed' )
+    
+    
     t_ctf_summed_preview = FileSensor( task_id='ctf_summed_preview',
         filepath="{{ dag_run.conf['directory'] }}/{{ dag_run.conf['base'] }}_ctf.jpg",
     )
@@ -312,9 +316,12 @@ dm2mrc {{ dag_run.conf['directory'] }}/{{ dag_run.conf['base'] }}-gain-ref.dm4 {
         jobid="{{ ti.xcom_pull( task_ids='convert_gain_ref' ) }}"
     )
 
+    t_influx_new_gainref = DummyOperator( task_id='influx_new_gainref' )
 
     t_wait_new_gainref = FileSensor( task_id='wait_for_new_gainref',
-        filepath="{{ dag_run.conf['directory'] }}/{{ dag_run.conf['base'] }}-gain-ref.mrc }}"
+        filepath="{{ dag_run.conf['directory'] }}/{{ dag_run.conf['base'] }}-gain-ref.mrc",
+        poke_interval=3,
+        timeout=60,
     )
 
 
@@ -354,11 +361,15 @@ MotionCor2  -InMrc {{ ti.xcom_pull( task_ids='wait_for_stack' ).pop(0) }} -OutMr
         jobid="{{ ti.xcom_pull( task_ids='motioncorr_stack' ) }}"
     )
 
+    t_influx_motioncorr_stack = DummyOperator( task_id='influx_motioncorr_stack' )
+
     t_motioncorr_2_logbbok = DummyOperator(task_id='upload_motioncorr_to_logbook')
 
 
     t_wait_for_aligned = FileSensor( task_id='wait_for_aligned',
         filepath="{{ dag_run.conf['directory'] }}/{{ dag_run.conf['base'] }}_stack.mrc }}",
+        poke_interval=5,
+        timeout=600,
     )
 
 
@@ -377,6 +388,8 @@ MotionCor2  -InMrc {{ ti.xcom_pull( task_ids='wait_for_stack' ).pop(0) }} -OutMr
         jobid="{{ ti.xcom_pull( task_ids='ctffind_stack' ) }}"
     )
     
+    t_influx_ctffind_stack = DummyOperator( task_id='influx_ctffind_stack' )
+    
     t_ctffind_stack_logbook = DummyOperator(task_id='upload_ctffind_to_logbook')
 
 
@@ -391,7 +404,7 @@ MotionCor2  -InMrc {{ ti.xcom_pull( task_ids='wait_for_stack' ).pop(0) }} -OutMr
     t_parameters >> t_param_influx 
 
     t_parameters >> t_ctf_summed >> t_wait_ctf_summed
-    
+    t_ctf_summed >> t_influx_ctf_summed
 
     t_ensure_slack_channel >> t_invite_to_slack_channel
     t_ensure_slack_channel >> t_slack_preview
@@ -408,11 +421,13 @@ MotionCor2  -InMrc {{ ti.xcom_pull( task_ids='wait_for_stack' ).pop(0) }} -OutMr
 
     t_wait_stack >> t_motioncorr_stack
     t_wait_gainref >> t_gain_ref >> t_wait_job_new_gainref >> t_motioncorr_stack
+    t_wait_job_new_gainref >> t_influx_new_gainref
     t_wait_new_gainref >> t_motioncorr_stack
     t_motioncorr_stack >> t_wait_motioncorr_stack 
+    t_wait_motioncorr_stack >> t_influx_motioncorr_stack
 
     t_wait_motioncorr_stack >> t_ctffind_stack
     t_wait_motioncorr_stack >> t_motioncorr_2_logbbok 
 
     t_wait_for_aligned >> t_ctffind_stack >> t_wait_ctffind_stack >> t_ctffind_stack_logbook 
-
+    t_wait_ctffind_stack >> t_influx_ctffind_stack
